--- a/packages/typescript-estree/src/ast-converter.ts
+++ b/packages/typescript-estree/src/ast-converter.ts
@@ -5,11 +5,11 @@
  * @copyright jQuery Foundation and other contributors, https://jquery.org/
  * MIT License
  */
-import convert, { getASTMaps, resetASTMaps, convertError } from './convert';
+import { convertError, Converter } from './convert';
 import { convertComments } from './convert-comments';
 import { convertTokens } from './node-utils';
 import ts from 'typescript';
-import { Extra } from './temp-types-based-on-js-source';
+import { Extra } from './parser-options';
 
 export default function astConverter(
   ast: ts.SourceFile,
@@ -27,16 +27,13 @@ export default function astConverter(
   /**
    * Recursively convert the TypeScript AST into an ESTree-compatible AST
    */
-  const estree: any = convert({
-    node: ast,
-    parent: null,
-    ast,
-    additionalOptions: {
-      errorOnUnknownASTType: extra.errorOnUnknownASTType || false,
-      useJSXTextNode: extra.useJSXTextNode || false,
-      shouldProvideParserServices
-    }
+  const instance = new Converter(ast, {
+    errorOnUnknownASTType: extra.errorOnUnknownASTType || false,
+    useJSXTextNode: extra.useJSXTextNode || false,
+    shouldProvideParserServices
   });
+
+  const estree = instance.convertProgram();
 
   /**
    * Optionally convert and include all tokens in the AST
@@ -52,11 +49,9 @@ export default function astConverter(
     estree.comments = convertComments(ast, extra.code);
   }
 
-  let astMaps = undefined;
-  if (shouldProvideParserServices) {
-    astMaps = getASTMaps();
-    resetASTMaps();
-  }
+  const astMaps = shouldProvideParserServices
+    ? instance.getASTMaps()
+    : undefined;
 
   return { estree, astMaps };
 }