--- a/packages/typescript-estree/src/ast-nodes.ts
+++ b/packages/typescript-estree/src/ast-nodes.ts
@@ -1,6 +1,6 @@
 import * as es from './ast-tree-nodes';
 
-export type ASTTreeNodes =
+export type ASTNode =
   | es.ArrayExpression
   | es.ArrayPattern
   | es.ArrowFunctionExpression