… packages/typescript-estree/src/estree/experimental.ts
… packages/typescript-estree/src/estree/extensions.ts
… packages/typescript-estree/src/estree/spec.ts
… packages/typescript-estree/src/parser-options.ts
… packages/typescript-estree/src/parser.ts
… packages/typescript-estree/src/temp-types-based-on-js-source.ts
… packages/typescript-estree/src/tsconfig-parser.ts
… packages/typescript-estree/src/typedefs.ts
… packages/typescript-estree/tests/lib/comments.ts
… packages/typescript-estree/tests/lib/javascript.ts
… packages/typescript-estree/tests/lib/jsx.ts
… packages/typescript-estree/tests/lib/parse.ts
… packages/typescript-estree/tests/lib/semanticInfo.ts
… packages/typescript-estree/tests/lib/tsx.ts
… packages/typescript-estree/tests/lib/typescript.ts
… packages/typescript-estree/tools/test-utils.ts