This is prototype of converter with types, with small improvement to performance (this is not goal of this PR).

Changes:
- type checking of AST structure
- reduces amount of calls for location data (location is calculated once per node)
- provides easier way to support range compiler flag (its not supported yet)
- remove old/unused estree definition files
- migrates ParserOptions from `temp-types-based-on-js-source.ts` to `parser-options.ts`

<details><summary>...</summary>

**before**
Complex File: `ts-estree x 238 ops/sec ±2.14% (86 runs sampled)`
Simple File: `ts-estree x 10,706 ops/sec ±3.19% (87 runs sampled)`

**after**
Complex File: `ts-estree x 262 ops/sec ±1.78% (84 runs sampled)`
Simple File: `ts-estree x 11,049 ops/sec ±3.26% (80 runs sampled)`

</details>