@bradzacher 2019-01-30
we should really raise a separate PR to remove all of the JSDoc types from the codebase...