@JamesHenry 2019-02-03
I manually ran some real-world "benchmarks" by building from this PR and including it in a build of Prettier and running it on a real codebase (angular.io documentation website) and can confirm there are not any performance regressions/issues